Export some function into services + finished addr/customer management
